# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Research Outreach Agent API
  version: 0.1.0
  description: GPT Actions schema for broad AI research opportunity discovery using OpenAlex.
servers:
  - url: https://YOUR_NGROK_OR_DEPLOYED_URL
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Check backend health
      responses:
        "200":
          description: Backend is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                      - ok
                required:
                  - status
  /version:
    get:
      operationId: getVersion
      summary: Confirm backend code version
      responses:
        "200":
          description: Backend code version.
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                    enum:
                      - 0.2-diverse-recent-search
                required:
                  - version
  /search_researchers:
    post:
      operationId: searchResearchers
      summary: Search for broad AI research opportunities and candidate researchers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchResearchersRequest"
      responses:
        "200":
          description: Candidate researchers and representative papers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResearchersResponse"
        "422":
          description: Invalid request.
        "502":
          description: OpenAlex request failed.
components:
  schemas:
    SearchResearchersRequest:
      type: object
      properties:
        topic:
          type: string
          minLength: 1
          description: Broad AI topic to search for.
          examples:
            - artificial intelligence
        limit:
          type: integer
          default: 5
          minimum: 1
          maximum: 50
          description: Maximum number of candidates to return.
        save_csv:
          type: boolean
          default: false
          description: Save returned candidates to outputs/candidates.csv.
        max_authors_per_paper:
          type: integer
          default: 1
          minimum: 1
          maximum: 5
          description: Maximum candidate authors to return from each paper.
        from_year:
          type: integer
          default: 2023
          minimum: 1900
          description: Only search papers published from this year onward.
        sort_by:
          type: string
          default: relevance
          enum:
            - relevance
            - citations
            - recent
          description: Sort OpenAlex results by relevance, citation count, or recency.
      required:
        - topic
    Candidate:
      type: object
      properties:
        source:
          type: string
          enum:
            - openalex
        paper_id:
          anyOf:
            - type: string
            - type: "null"
        name:
          type: string
        institution:
          anyOf:
            - type: string
            - type: "null"
        author_url:
          anyOf:
            - type: string
            - type: "null"
        paper_title:
          anyOf:
            - type: string
            - type: "null"
        paper_year:
          anyOf:
            - type: integer
            - type: "null"
        paper_url:
          anyOf:
            - type: string
            - type: "null"
        abstract:
          anyOf:
            - type: string
            - type: "null"
        cited_by_count:
          anyOf:
            - type: integer
            - type: "null"
        opportunity_angle:
          type: string
        lab_signal_score:
          type: number
          minimum: 0
          maximum: 1
        paper_signal_score:
          type: number
          minimum: 0
          maximum: 1
        project_activity_score:
          type: number
          minimum: 0
          maximum: 1
        contactability_score:
          type: number
          minimum: 0
          maximum: 1
        contribution_angle_score:
          type: number
          minimum: 0
          maximum: 1
        profile_fit_score:
          type: number
          minimum: 0
          maximum: 1
        search_signal_score:
          type: number
          minimum: 0
          maximum: 1
        fit_score:
          type: number
          minimum: 0
          maximum: 1
        contact_priority:
          type: string
          enum:
            - high
            - medium
            - low
        suggested_outreach_angle:
          type: string
      required:
        - source
        - paper_id
        - name
        - institution
        - author_url
        - paper_title
        - paper_year
        - paper_url
        - abstract
        - cited_by_count
        - opportunity_angle
        - lab_signal_score
        - paper_signal_score
        - project_activity_score
        - contactability_score
        - contribution_angle_score
        - profile_fit_score
        - search_signal_score
        - fit_score
        - contact_priority
        - suggested_outreach_angle
    SearchResearchersResponse:
      type: object
      properties:
        topic:
          type: string
        limit:
          type: integer
        count:
          type: integer
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
        csv_path:
          anyOf:
            - type: string
            - type: "null"
      required:
        - topic
        - limit
        - count
        - candidates